moved src/app/api/lib -> src/lib/be
